Improve Model-View pattern and add Resource Manager.
Now, elements of the Model should inherit from ModelElement (which forces them to be JSON-Parseables), and views must inherit from View, which forces them to initialise binding to an element of the model.

Also a class to manage different sprites across tiles allows for better allocation of configuration, instead of configuring each individual tile.