Modify how map elements are loaded from the JSON.
Also modifies the order of loading and switchs from constructor-initialisation to lazy initialisation. This is due to the need for accessing the list of resources of the map while we are loading the map elements. If it is done in the constructor, the reference to the game and the map is still not being assigned to the singleton instance of the GameManager, therefor, trying to access the resources using GameManager.Game.Scenario.Map.Resources will crash the game, as GameManager.Game is not still assigned (we are in the constructor of Game!!!).